ref: combine the two auth tests
